Issue-408: WIP Enhancement: Refactor Signup Page into smaller chunks
