+++ src/components/Signup/index.js
@@ -1,0 +1,14 @@
+// pod helpers
+import PodRegistrationForm from './pod/Signup';
+
+/* Components */
+import { initializePod, registerPod } from './Components/PodRegistrationForm';
+import ShowNewPod from './Components/ShowNewPod';
+
+/**
+ * Components and functions related to Signup functionality within project PASS
+ *
+ * @namespace Signup
+ */
+
+export { PodRegistrationForm, ShowNewPod, initializePod, registerPod };
